Translation setup
Setup translation for this repo as django project following the
common setup as documented in
http://specs.openstack.org/openstack-infra/infra-specs/specs/translation_setup.html

This adds the django babel files and removes the default one which will
not work with django.

Fixed package name, all python files are in ironic_ui, so package name
needs to be ironic_ui, not ironic-ui.

Change-Id: I8482a1896f16b4ed19a1117c7afc583bba5e13fb