Added success messages to power change
When power change occurs successfully notify user that this has
happened and they need to refresh page to see the updated power status
in the node details.

Currently, prior to this patch, an updated power state will result in a
blank field where the power state would normally show on/off.

Change-Id: Ibbde2bc60a6f3c9778e62b0de61c70d27d1189f4