Changed dropdown menu to align right on driver details page
When selecting an image on the driver details page of 'enroll node'
the dropdown menu was aligned to the left of the button meaning that
a lot of the menu and it's content was cut off. I have re-aligned
this menu to the right so that the full dropdown is viewable and not
cut off at the edge of the modal.

Change-Id: Icff5eab3d0cae82bc8d69103bd7ee0a46476c7c9