Consolidate node last_error processing
Consolidate last_error processing for the node-list and node-detail
views using a newly added node-error service. The service provides
a function checkNodeError that will compare the error condition
of a specified node with its last known value which is stored
in browser session persistent storage (as part of each check).
If a change in the node error condition has occurred the user
is notified using the toast service. The use of persistent storage
eliminates duplicate notifications that would otherwise be
generated from page transitions and service re-initializations.

The checkNodeError function is called whenever node information
is retrieved using the ironic service getNode/getNodes functions.

Change-Id: Iba7bdaaa78d51384b6b9d79d2d723b8e7607eb9a