Extend support for the Ironic state machine
The following changes have been made:
(1) Create a separate node-state-transition service that is used
to orchestrate changes in the node provision state
(2) Extend the node list and detail UIs (actions) to support
states other the manageable, available, and active
(3) It should be noted that state transition actions are rendered
as html list items as opposed to actions. The reason for this is
that the <action> tag is not compatible with the ng-repeat
directive.

Change-Id: Id6543667f944c47f8f9727c1c2f3085ae75a1625