Create base-port module to support create and update operations
The base-port module was created by refactoring the existing
create-port module, and will be used to support both port
create and update operations. The base-port module contains
an html form for the port and a controller that is used
to supervize property modifications. Controllers for the create
and update operations inherit from the base controller.

Change-Id: I0749241a65a7db130f94cad180afb30ff9e578ba