Fix node validation problem
Fixed an issue in the display of node validation information
that cause a javascript runtime error. The fix is in two parts:
(1) removed an empty sucess processing function in the
ironic.service/validateNode function that caused the response to
undefined to callers, and (2) I have found that table widget does
not always update correctly when the data is an object. The
base validation data is now provided as an array.

Change-Id: I0e9c0e0c953d0fbd5d4732aa32975a30933e4a6f