Cleanup code associated with the ironic REST api
- In the ironic_rest_api module use specific names like node, port,
driver for return values. I believe this is more descriptive than
the current approach of using "items" in all function returns. The
use of specific names also reduces the possibility of cut and paste
errors in client code.
- Rethink data returned from functions in the ironic.service. For
example, I believe that a function like getNodes should return a
list of nodes as opposed to a HTTP response as is done today. The
client should receieve the requested data, and be abstracted from
how it was obtained
- Modify controllers to take advantage of the api simplifications

Change-Id: Ica8c9ff8dfa9299cd5ef2893d296e366065bca97