Add support for starting/stopping/accessing the serial console
- The action menu in the node-details panel has a new item
Enable|Disable console.
- The node-details/overview panel has a new "Console info." item
in the "General" section. The value of this field is dependent on
the console type. For "shellinabox" the value is an anchor with
the URL of the console and title "shellinabox". For others, the
value is a string representation of the console_info object
returned by the get_console api call.

Closes-Bug: #1655690
Change-Id: Ibfa687f50d2251987276367d743ffb2a88e52796