Fix Javascript unassignable errors
Replaced several instances of the action directive with HTML list-item/
anchor equivalents. Because the action uses two-way bound scope
variables for "callback" and "item" arguments you cannot specify a
literal value as an argument to the callback function. Two-way bound
variables must be assignable. In my mind there is some question as
to whether the action directive should have been implemented in this
way, but for now I am proposing to use an ng-click directive within
an anchor to implement callback functionality when a literal argument
is needed.

Change-Id: I2baf17c2784faeb92da362544c53e444e6d01afc