Unit test framework for Ironic-UI API service
This commit is a first step in the development of a unit test framework for
the Ironic-UI API service. The approach being used is to provide an emulation
of the Ironic backend using a mock that utilizes Angular $httpbackend
handlers to intercept requests targeted at the Ironic-UI server side REST
endpoints.

In addition to demonstrating the mock I have coded some basic tests to
show how it is used. At this point I would like to get feedback from the team,
and consensus on whether we should continue the process of building
additional tests.

Change-Id: Iaec83b0e19b5051ebf1257ddc56efcc6f11ee39d