Introduce form-field directive
The form-field directive abstracts the data, presentation, and operation of
a form field into a reusable component. It is intended to improve
maintainability and consistency. Form-field is currently being used to
create and edit ports, however, the goal is to use it throughout the
Ironic-UI. Componentizing form elements will also make it easier to
move to other libraries such as SchemaForm should we decide to do that.

Change-Id: I4cd84cd0840878b3dc34afe9def8e7f94be46fd0