Fix horizon dependency
This commit moves horizon to requirements.txt.

horizon was added to upper-constraints in stable branches recently,
so installing from the tarball no longer works. upper-constraints
in stable branches picks up the right version of horizon, so there
is no need for the upper bound of the horizon version.

In addition, the link of the horizon tarball was not updated after
the stable branch was created. I don't know why it works for stable/stein
but I believe this commit mitigates the situation as the upper-constraints
file in the requirements repo picks up the right version of horizon.

This commit is proposed directly as the similar change in the master
branch happened as part of Django 1.11 drop [1].

In addition, ironic-ui tox.ini specified install_command with the
upper-constraints.txt. The contraints file is now specified in "deps"
section to follow the convention used widely in the OpenStack community.
doc/requirements.txt is cleanup and the runtime and test requirements
are now added to "deps" of the "docs" env to generate the module index
properly.

stable/stein only:
sphinx entry in test-requirements.txt did not match global-requirements.
This commit updates it to match the current global-requirements.

[1] https://review.opendev.org/#/c/700845/

Change-Id: I9a310446743cfe6bca5480455b91a7cd74da8785
(cherry picked from commit efa4927bb7baeb918fad3734b21b0223e0958404)