Fix horizon dependency
This commit moves horizon to requirements.txt.

horizon was added to upper-constraints in stable branches recently,
so installing from the tarball no longer works. upper-constraints
in stable branches picks up the right version of horizon, so there
is no need for the upper bound of the horizon version.

In addition, the link of the horizon tarball was not updated after
the stable branch was created. I don't know why it works for stable/train
but I believe this commit mitigates the situation as the upper-constraints
file in the requirements repo picks up the right version of horizon.

This commit is proposed directly as the similar change in the master
branch happened as part of Django 1.11 drop [1].

The way to specify upper-constraints.txt is also update to follow the
convention used widely in the OpenStack community. The contraints file is
now specified in "deps" section. This allows us to use the lower-constraints
file correctly. Previously upper-constraints.txt was actually used in the
lower-constraints job unexpectedly.

sphinx entry in test-requirements.txt did not match global-requirements.
This commit updates it to match the current global-requirements.

[1] https://review.opendev.org/#/c/700845/

Change-Id: I9a310446743cfe6bca5480455b91a7cd74da8785
(cherry picked from commit efa4927bb7baeb918fad3734b21b0223e0958404)
(cherry picked from commit 5c593fd34a149c62e4153a1456bb5bc1ae366ea7)