Stop configuring install_command in tox.
Currently, we are overriding 'install_command' to use 'pip'. This is
considered poor behavior and 'python -m pip' should be used instead:

https://snarky.ca/why-you-should-use-python-m-pip/

It turns out that this is the the default value provided by tox:

https://tox.readthedocs.io/en/latest/config.html#conf-install_command

So we can remove the line and simply use the default value.

openstack-discuss thread
http://lists.openstack.org/pipermail/openstack-discuss/2020-April/014237.html

Bumping the following libs in lower-constraints to avoid version
conflict:
-osprofiler 2.3.0
-python-keystoneclient to 3.22.0
-python-cinderclient to 5.0.0

Change-Id: Id0049dea1e16f9c459e7e63e5f7d96666fff708a