Remove Chrome browser from karma.conf.js
This patch removes Chrome browser from
karma.conf.js as nodejs job fails with Chrome
browser because it is not able to access some URL.
So let's remove Chrome browser now then we can add
it again once the issue is resolved.

Also as discussed during the upstream ironic community meeting on
Monday Dec 14 2020, the lower-constraints job is being removed.

Co-Authored-By: Riccardo Pittau <elfosardo@gmail.com>
Change-Id: Ic7094ab6ed01af2bcbe0beb13f300499215fd903